# Speedup stedc_mergeValues_kernel