The PR speeds up "merge Values" stage of the stedc for big cases (n > 1024) by splitting the stedc_mergeValues_kernel into multiple kernels and thus allowing to solve secular equations with full parallelism (original code had a limitation on the number of groups). Smaller merges are still done using nonsplitted stedc_mergeValues_kernel to avoid overhead.

The PR doesn't change algorithm or numerical properties of calculations and the output of the stedc is supposed to be bit exact to the previous implementation. However, only basic tests were done and PR might need additional testing and verification.