@Kirpich30000 2025-07-26
Here are a few timing results for the whole syevd gathered using rocsolver-bench

Size | Base time, us | PR time, us | Speedup
-- | -- | -- | --
1024 | 57147 | 53369 | 1.07
2048 | 143276 | 116694 | 1.23
4096 | 468226 | 347890 | 1.35
8192 | 2166989 | 1650779 | 1.31